Order BibTeX entries by key
This is for defining a stable-ish sort order so that a .bib file in version control doesn't change too much when adding new references.

The sorting happens in write.bibtex(), I assume that sorting won't break or change the meaning of a BibTeX file. (Duplicate keys: order() is a stable sort and won't change their order.) However, for my use case it would be sufficient to reorder in get_bib().